refactor: move computation of manifest hash to manifest crate
