+++ crates/proof-impl/checkpoint-new/Cargo.toml
@@ -9,6 +9,7 @@ workspace = true
 [dependencies]
 ssz.workspace = true
 ssz_primitives.workspace = true
+strata-asm-manifest-types.workspace = true
 strata-checkpoint-types-ssz.workspace = true
 strata-crypto.workspace = true
 strata-ledger-types.workspace = true
